create state.js, button.js and move functions into their own files, import at origin
